Added "mini.sessions"
Session management is actually very, very handy and useful for somewhat
large projects. Also if you quit Neovim to reopen it pretty often. It
keeps track of the currently open buffers and stuff it's really nice.
Just make sure to create a session with ":mksession" command and then
"mini.sessions" will handle automatically reading and saving your
session. Also disabled ghost text for completions because it can get
really annoying if the completion is over a line long so it jumps the
text/code around while typing which is just, annoying. You can turn it
back on if you want of course if you prefer.